Simplify ListSerializer to use a single hook
ListSerializer really only needs a single hook that is invoked before
each value is written, and by making this hook an IntComsumer that
can inspect if the first value or subsequent values are being written,
there's no need for a between-values hook. To know if any values were
written, the ListSerializer now exposes its position which can be
queried after values are serialized.