Use ShapeSerializer to write structures
Rather than introduce a dedicated interface for writing struct/union,
we can just reuse ShapeSerializer. This simplifies implementing and
handling parts of the data model since before we often had to implement
both a StructSerializer and ShapeSerializer to truly intercept all
writes for a structure (there was some shape delegation that had to
be intercepted too).

Along with this change, the "stringMemberIf" kinds of methods are now
static helper methods on ShapeSerializer. The code generator will
need to account for this.

ShapeSerializer#flush no longer is defined as throwing an IOException
and is implemented by default since most implementations don't need
to be flushed.