Add map key schema, writeNull, cleanup SdkSchema
MapSerializer now sends the key schema when writing an entry to avoid
needing to look it up each time an entry is serialized. The method
entry was also renamed to writeEntry to match the other serialization
methods.

Added writeNull to ShapeSerializer to explicitly capture null value
serialization. This is used specifically to serialize null values in
sparse lists and maps.

Various cleanup was made to SdkSchema.