Refactor memberIndex, documents, and maps
This commit refactors memberIndex so that it's more clear that the
index is just the index position a member is in the result of calling
members() on a schema. memberIndex is no longer set when creating a
member, but rather when a member is added to a containing shape. This
removes the ability to set custom indices and better correlates the
index to the position in members().

When creating the map of members and list of members for a schema,
the SdkSchema class now creates the most tailor-made Map implementation
based on the shape type and member count (e.g., a list suppports a
single member named member, so ListMembers provides that validation
and simplified implementation). This is for both performance, memory,
and validation.

This commit removes integer and long map key support. We don't actually
support these kinds of maps yet and can add support for them later
if/when we add support. There's lots of refactoring coming up, and
removing these now prevents future churn in PRs that'll just get
removed until generalized support is added.

Finally, this commit refactors documents so that schemas are captured
in documents, making it far simpler to implement typed documents.
Typed document deconstruction (creating documents from a typed
POJO) is still deferred until if/when access to the typed document is
requested. This makes it very quick to simply wrap a type and serialize
it as a document, but also makes it easier to understand how a document
is deconstructed into nodes.