Improve documents, JSON, and timestamp formatter
This commit adds fairly extensive test coverage to the JSON codec,
and minor performance improvments around things like reading and
writing member names that may or may not take jsonName traits into
account. Various bug fixes were made along the way.

Codec, ShapeSerializer, and ShapeDeserializer now implement AutoClosable
so that they can potentially support caching in the future (there's a
TODO hook for us to support caching of byte buffers for the JSON codec).

Some cleanup was done to TimestampFormatter to improve the API and
couple the "of" method to known formats from the TimestampFormatTrait.

A "normalize" method was added to Document to take a document and
create a normalized version of it that has no schema or protocol
details, making it easier to compare documents.