Add __type to structure documents (#139)
Typed structure documents will return the shape of the captured shape
when a synthetic __type member is accessed.

The JSON codec has been updated to serialize the shape ID using
__type. Most protocols will follow this same pattern, though it's
theoretically possible that a future codec may wish to deviate.

This commit also updates lazy document handling so that only structures
are lazily converted to documents. This ensures that
shape -> document -> serialized actually is handled without having
to create a DOM version of a structure just to serialize it.

When serializing nested document values within a document,
Document#serialize is now used rather than Document#serializeContents.
This gives codecs the opportunity to add any necessary framing to
structure documents, and this method delegates to serializeContents.