Make TypeRegistry an interface
An interface allows for more abstraction over other potential ways to
match shape IDs to builders. The registry now returns null instead
of Optional values. Updated codegen to only register errors with
operation registries as those are the only place they're being used.

A followup is to generate service-wide errors into the registry
of a client, and merge it with the operation registry. Further merging
would be to support framework errors and potentially synthetic errors.