Context: remove keys(), add putAll, copy, view methods (#233)
Remove `.keys()` method that allowed iteration over all keys of a
Context, to prevent someone who doesn't have Java visibility to the
Context.Key definition from being able to access that key's
value. e.g., If a Context.Key is package-private, then only classes
from the same Java package can access the value of the key from a
Context.

`.keys()` was mainly there to support being able to copy/add a
Context. So instead adding a `putAll()` to Context. Made Context a sealed
interface that doesn't require exposing `keys()` on the interface.

Also, can get an unmodifiable "view" of the Context, that can be
passed to components where we want the guarantee that they can't
modify it. Note, this has "view" semantics, where changes to the
underlying Context would be visible via the unmodifiable view.

Also added `unmodifableCopy` and `modifiableCopy` methods to Context.