Revert 286 to make JsonFieldMapper an interface
To void contention by using a global cache of JSON names across all
clients, this commit reverts #286. Implementations that can optimize
getting member names of structures when using member names can now
just check if the fieldMapper is == to UseMemberName.INSTANCE since
the only instance of UseMemberName can be the INSTANCE.

This commit also updates the caches so that they actually populate.