Add direct traits, fix member serde, finish XML
This commit finishes implementing rest-xml and the XML codec. This
required adding "direct traits" and fixing member serde.

With the XML codec, we need to sometimes check if a trait is applied
directly to a member and not use traits that are applied to the shape
targeted by the member. This change adds support for this by adding
"getDirectTrait" to Schema.

When serializing XML codecs, it necessary to be very careful about
whether we send a member schema or a root shape schema to serializers.
We previously were not using map key members, map value members, or
list members when serializing these shapes. However, this doesn't work
with XML codecs because these members can have traits on them like
XML namespaces. This change adds getters to Schema to get these members
directly without needing to do any kind of hash based looked on member
names.