Add schema to ShapeBuilder to aid http bindings
ShapeBuilder now provides a method that returns the Schema of the
shape to build. This allows things like HTTP bindings to have more
prior knowledge about what's about to be deserialized so they can more
efficiently deserialize. In this case, HTTP binding's BindingMatcher
now uses an array-based cache that indexes into it using the
memberIndex of each member of the struct to match against. This helps
with the repeated binding checks that are done on each
request and response.

In addition to the above, several minor issues and extraneous methods
were fixed or removed (e.g., RequestDeserializer never needs to
deserialize an error).