Use ClassValue based TraitMap for perf
The previous TraitMap approach showed well in previous benchmarks
due to optimizations that made the benchmark more unreliable. This
PR moves TraitMap to a single implementation that's based on
ClassValue, a class used to attach metadata to classes. In this case,
we assign each class an auto-incremented ID and then use that as the
"universe" of possible traits for a schema, allowing us to directly
index into an array of traits. The previous approach resulted in
megamoprhic call sites that were slower in practice than using just
a single ClassValue-based implementation. This is also surprisingly
faster than a HashMap of Class to Trait, a linear search, binary
search, uniform binary search, custom probing hashmaps, etc.