Switch to Map storage when more than 64 keys
We also need to change from Context.putAll which gets values from a
Context using internal class access, to using Context.copyTo to
publish from one context to another.

When the keyspace exceeds 64 keys, future contexts are created using
a map-based storage, reducing the memory overhead of creating contexts.
Previously created array-based contexts might still receive values
from a map context, and that's fine, they'll grow to account for it.
However, array contexts might need to later copy their contents to
another context, and in that case, they need to also hold onto the
keys provided for each value. To limit the overhead of this, keys
that fit into the array keyspace are stored centrally on Context.Key.
Keys outside this keyspace added to an array-based context will store
the key in a map so they can be recalled later if copyTo is used.