Deserialize generated types with member schemas
Some protocols, like XML, apply traits to members that target
structures, enums, etc, and those traits need to be accounted for
during deserialization. This wasn't possible before, and made it so
that we couldn't fully implement the XML codec. This PR updates
ShapeBuilder to add a new method for deserializing a shape that also
accepts a Schema to provide it access to member schema traits.