Add DynamicClient to use without codegen
The DynamicClient can be used when you need to interact with
services more dynamically and can't code-generate a client for
every service. It can dynamically load Smithy models at runtime,
convert them to a schema-based client, and call the service using
document types as input and output. The client currently does not
support streaming or event streaming.