Add check for implicit conflicts (#412)
Adds a check for potential implicit name duplication between a generated type in the same package and a base type. For example `IllegalArgumentException`'s are used within the generated code (in enums and unions), but will conflict with any generated type named `IllegalArgumentException`, so we need to use the fully qualified name for the java.lang illegal arg exception.