Replace exception with result value for identity
Rather than throw an exception when an identity cannot be resolved,
this commit now has us return an IdentityResult<T>. This makes it so
that we don't have to create a bunch of exceptions and their stack
traces every time we run through a chain of identity resolvers.

IdentityResult is meant to capture expected errors, like missing
environment variables or missing config files. It is not meant to
capture unexpected errors like malformed input, corrupted files, or
networking errors. Just throw exceptions for those.