Integrate retries with the pipeline
Retries are now integrated into ClientPipeline so that retries are
actually performed.

To achieve this, ApiException now contains mutable properties used
to indicate the retry information of an error. Interceptors and
protocols can mutate these values as needed before ultimately handing
it off to a RetryStrategy to make the retry decision.

To account for schema based retry information, the schema of an
error is now available via schema() on ModeledApiException.

Retry requests and responses have been simplified, and ClientCall
was simplified into a POJO with package-private properties.