Remove ability to change transport per/request
We'll allow a transport to be configured only at client creation.
A protocol can still be swapped out per/request. Transports bring
with them potentially a lot of setup, like connection pools, so they
have to do more work up front to configure themselves with a client.
If a transport needs to change, then a new client needs to be created.