Rename ApiException and ModeledApiException and move
ApiException is renamed to CallException to better indicate that the
error may not have actually hit a service or be related to the API
(e.g., it could be a timeout, networking error, protocol error, etc.)

ModeledApiException is renamed to ModeledException since it no longer
extends from ApiException and Api isn't helping make the name more
clear.

Exceptions have been moved to a top-level error package, since not all
errors are modeled or have a schema, and to better group them.