Add initial implementation of synchronous Waiters (#571)
Adds initial implementation of Synchronous waiters.

Waiters are a client-side abstraction used to poll a resource until a desired state is reached, or until it is determined that the resource will never enter into the desired state. Waiters can be created without codegen, but a codegen integration is provided to automatically create waiters base on trait in the service model.