Add a way to register classes without eager load
When registering a lot of classes in a type registry, you don't want to
eagerly load all the classes up front. This would pull in virtually the
entire graph of the model at startup.

This adds an alternative, albeit more cumbersome, API to register
classes and builders with a type registry that only eagerly loads
classes on first use by putting them into Suppliers.