Produce service description bundles for AWS services
This commit adds the Bundler interface and an implementation for AWS
services. The AWS service bundler will download the specified service's
Smithy model from the public AWS model repository, extract its
endpoints, remove some unnecessary elements from its model, and package
it up into a Bundle. The smithy-mcp server will then load this bundle,
parse it, and use the AwsAuthProvider defined in the aws-mcp-provider
module to dynamically serve the bundle's model as a set of MCP tools and
appropriately configure calls to that service.