Add payloadCodec to ClientProtocol
ClientProtocol now exposes a payloadCoded method that can be used to
get the Codec configured for the protocol. So a RestJson1 protocol
would be able to return Codec configured to correctly serialize and
deserialize RestJSON1 payloads. If a protocol ever comes along that
has no payload codec or perhaps multiple, null may be returned.