Add ability to override call config in interceptor
Interceptors now have a new hook, modifyBeforeCall that accepts the
ApiOperation being called, input of the call, and ClientConfig. It
is expected to return the same ClientConfig or an updated ClientConfig
using ClientConfig#withRequestOverride.

This allows for much more dynamic calls with clients, like changing
the protocol per/call, applying plugins per/call, and more.