Make Document#discriminator nullable, add registry
Document#discriminator is now nullable, and returns null if no
discriminator is found. This method still throws if a string is found
but cannot be parsed. Making this nullable allows us to compose
type registries so that they can check multiple registries to attempt
to create a builder for a document.

Added a new api to TypeRegistry: createBuilder(Document). This API
attempts to find a builder for a document based on its discriminator
or returns null if the document has none, or no matching builder is
found. This matches the behavior of createBuilder(ShapeId).

Added a new api to TypeRegistry: deserializeString(Document). This API
deserializes a document into a builder from the registry but does not
use error correction, ensuring that missing defaults are not filled in
with zero values, and deserialization fails when members are missing
(predicated on the resolved builder performing the required-ness check).