Adding support for Prompts in the MCP server.
The prompts have a `prompts/list` that allows tools such as Q CLI to call /prompts and get a list of prompts. The `prompts/get` allows getting and invoking the prompt with arguments. Q CLI uses `@<promptName> [args]` to define which prompt is invoked.

Prompts can be provided via a configuration file.